+++ Faker/Helper/YamlLoader.cs
@@ -1,4 +1,5 @@
 ﻿using RimuTec.Faker.Extensions;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -44,30 +45,73 @@ namespace RimuTec.Faker.Helper
       internal static string Fetch(string locator)
       {
          var localeName = Config.Locale;
-         if (Config.Locale == "en")
-         {
-            localeName = $"{localeName}.{locator.Split('.')[0]}";
-         }
 
          // if locale hasn't been loaded yet, now is a good time
-         YamlNode fakerNode;
-         if (!_dictionary.ContainsKey(localeName))
-         {
-            using (var reader = YamlLoader.OpenText($"RimuTec.Faker.locales.{localeName}.yml"))
-            {
-               var stream = new YamlStream();
-               stream.Load(reader);
-               YamlNode rootNode = stream.Documents[0].RootNode;
-               var localeNode = rootNode[Config.Locale];
-               fakerNode = localeNode["faker"];
-               _dictionary.Add(localeName, fakerNode);
-            }
-         }
+         LoadLocale(localeName, locator);
 
          // at this point the locale is in the dictionary
-         fakerNode = _dictionary[localeName];
-         var locatorParts = locator.Split('.');
-         return Fetch(fakerNode[locatorParts[0]], locatorParts.Skip(1).ToArray());
+         YamlNode fakerNode;
+         var key = localeName;
+         try
+         {
+            fakerNode = _dictionary[key];
+            var locatorParts = locator.Split('.');
+            return Fetch(fakerNode[locatorParts[0]], locatorParts.Skip(1).ToArray());
+         }
+         catch {
+            // fall back to locale "en"
+            LoadLocale("en", locator);
+            key = $"en.{locator.Split('.')[0]}";
+            fakerNode = _dictionary[key];
+            var locatorParts = locator.Split('.');
+            return Fetch(fakerNode[locatorParts[0]], locatorParts.Skip(1).ToArray());
+         }
+      }
+
+      private static void LoadLocale(string localeName, string locator)
+      {
+         var key = localeName;
+         if(localeName == "en")
+         {
+            key = $"{localeName}.{locator.Split('.')[0]}";
+         }
+
+         if (!_dictionary.ContainsKey(key))
+         {
+            YamlNode fakerNode;
+            var executingAssembly = Assembly.GetExecutingAssembly();
+            string embeddedResourceFileName = $"RimuTec.Faker.locales.{key}.yml";
+            if (executingAssembly.GetManifestResourceNames().Contains(embeddedResourceFileName))
+            {
+               using (var reader = OpenText(embeddedResourceFileName))
+               {
+                  var stream = new YamlStream();
+                  stream.Load(reader);
+                  YamlNode rootNode = stream.Documents[0].RootNode;
+                  var localeNode = rootNode[localeName];
+                  fakerNode = localeNode["faker"];
+                  _dictionary.Add(key, fakerNode);
+               }
+            }
+            else
+            {
+               var assemblyLocation = new FileInfo(executingAssembly.Location);
+               var fileName = Path.Combine(assemblyLocation.DirectoryName, $"{key}.yml");
+               if (File.Exists(fileName))
+               {
+                  var yamlContent = File.ReadAllText(fileName);
+                  using (var reader = new StringReader(yamlContent))
+                  {
+                     var stream = new YamlStream();
+                     stream.Load(reader);
+                     YamlNode rootNode = stream.Documents[0].RootNode;
+                     var localeNode = rootNode[localeName];
+                     fakerNode = localeNode["faker"];
+                     _dictionary.Add(key, fakerNode);
+                  }
+               }
+            }
+         }
       }
 
       private static string Fetch(YamlNode yamlNode, string[] locatorParts)